login page 및 card 컴포넌트 코드 수정
