+++ src/pages/LoginPage.tsx
@@ -1,14 +1,97 @@
-import React from "react";
+import React, { useState } from "react";
 import { googleIcon, loginPageCharatorImg } from "../assets/images";
 import InputField from "../components/loginpage/InputField";
 import Button from "../components/loginpage/Button";
 import Card from "../components/loginpage/Card";
+import Header from "../components/loginpage/Header";
+import { useNavigate } from "react-router-dom";
 
 const LoginPage: React.FC = () => {
+  const [loginForm, setLoginForm] = useState({
+    email: "",
+    password: "",
+  });
+
+  const navigate = useNavigate();
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>): void => {
+    e.preventDefault();
+  };
+  const handlePssword = () => {
+    navigate("/");
+  };
+  const handleGoogle = () => {
+    navigate("/");
+  };
+  const handleSignInForm = (e: React.ChangeEvent<HTMLInputElement>): void => {
+    const { type, value } = e.target;
+    setLoginForm({ ...loginForm, [type]: value });
+  };
   return (
-    <div className="w-full h-screen flex justify-center  bg-gradient-to-r from-sky-300 to-pink-500 items-center   ">
-      <Card />
-    </div>
+    <Card>
+      <section className="w-4/6  h-full flex flex-col p-6 max-lg:w-full">
+        <Header logo="LoGo Here" text="Welcome home" />
+        <form
+          className="flex flex-col justify-between max-lg:items-center w-3/4 max-lg:w-full"
+          onSubmit={handleSubmit}
+        >
+          <InputField
+            id="user_id"
+            placeholder="Email"
+            type="email"
+            onChange={(e) => handleSignInForm(e)}
+            value={loginForm.email}
+          />
+          <InputField
+            id="user_password"
+            placeholder="Password"
+            type="password"
+            error="패스워드가 틀렸습니다."
+            onChange={(e) => handleSignInForm(e)}
+            value={loginForm.password}
+          />
+          <p
+            className="mb-4 text-xs flex items-center text-slate-400 hover:text-black cursor-pointer justify-center"
+            onClick={handlePssword}
+          >
+            For got password?
+          </p>
+          <div className="w-full flex justify-center">
+            <Button customStyle="bg-pink-400" type="submit">
+              Login →
+            </Button>
+          </div>
+        </form>
+        <section className="w-3/4 flex flex-col items-center max-lg:w-full max-lg:justify-center">
+          <p className="text-slate-400 text-xs mb-4">or continue with</p>
+          <Button
+            customStyle="flex items-center justify-center"
+            type="button"
+            onClick={handleGoogle}
+          >
+            <img
+              alt="구글아이콘"
+              src={googleIcon}
+              className="w-10 h-6 object-cover"
+            />
+          </Button>
+          <div className="flex justify-between items-center w-full">
+            <p className="text-slate-400 text-xs">
+              Don't have an account yet ?
+            </p>
+            <button className="text-slate-400 hover:text-black cursor-pointer  ">
+              Sign Up for free
+            </button>
+          </div>
+        </section>
+      </section>
+      <section className="w-2/6 max-lg:w-full max-lg:h-96 rounded-l-3xl bg-sky-100 flex items-center relative max-lg:hidden  ">
+        <img
+          src={loginPageCharatorImg}
+          alt="이미지"
+          className=" absolute h-full w-full object-cover right-44 mix-blend-darken"
+        />
+      </section>
+    </Card>
   );
 };
