+++ src/pages/SignUpPage.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import Card from "../components/loginpage/Card";
+
+const SignUpPage: React.FC = () => {
+  return <div></div>;
+};
+
+export default SignUpPage;
